added ability to fetch personal accounts count

+++ personalAccountsService/src/main/java/com/statkovit/personalAccountsService/payload/PersonalAccountFolderDto.java
@@ -27,5 +27,5 @@ public class PersonalAccountFolderDto {
     @Size(max = PersonalAccountFolder.MAX_LENGTH__NAME)
     private String name;
 
-    private Integer accountsCount;
+    private Long accountsCount;
 }
